// src/types/project.ts
export interface Project {
  title: string;
  description: string;
  tech: string[];
  image: string;
  github?: string;
  demo?: string;
  featured: boolean;
  year: string;
  slug: string;
  overview?: string;
  role?: string;
  duration?: string;
  challenges?: string[];
  solutions?: string[];
  features?: string[];
  gallery?: string[];
  architecture?: string;
  database?: string;
  lessons?: string[];
}